+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,8 +1,13 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import roomescape.dao.ReservationDAO;
+import roomescape.dto.ReservationRequestDTO;
+import roomescape.dao.ReservationTimeDAO;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 import java.util.List;
