+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,8 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import roomescape.dao.ReservationTimeDAO;
+import roomescape.domain.ReservationTime;
 
 import java.util.List;
