+++ src/main/java/roomescape/dao/ReservationTimeDAO.java
@@ -1,13 +1,13 @@
-package roomescape;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
 
 import java.sql.PreparedStatement;
-import java.sql.Time;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Objects;
