+++ src/test/java/roomescape/dao/ReservationTimeDAOTest.java
@@ -1,10 +1,11 @@
-package roomescape;
+package roomescape.dao;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
 import java.util.List;
